Use Fody.Costura to merge generated assembly for Papercut.Module.WebUI; Fixed issues in AssemblyScanner when a plugin has a customized reference other than pre-referenced assemblies; Renamed Papercut.WebUI projects to Papercut.Module.WebUI projects
