+++ appveyor.yml
@@ -10,6 +10,7 @@ after_build:
   - cd %APPVEYOR_BUILD_FOLDER%\src\Papercut.UI\bin\Release
   - md plugins
   - move Papercut.Module.*.dll plugins
+  - move Papercut.Module.*.config plugins
   - 7z a Papercut.%APPVEYOR_BUILD_VERSION%.zip plugins
   - 7z a Papercut.%APPVEYOR_BUILD_VERSION%.zip Papercut.exe
   - 7z a Papercut.%APPVEYOR_BUILD_VERSION%.zip Papercut.exe.*
@@ -19,6 +20,7 @@ after_build:
   - cd %APPVEYOR_BUILD_FOLDER%\src\Papercut.Service\bin\Release
   - md plugins
   - move Papercut.Module.*.dll plugins
+  - move Papercut.Module.*.config plugins
   - 7z a PapercutService.%APPVEYOR_BUILD_VERSION%.zip plugins  
   - 7z a PapercutService.%APPVEYOR_BUILD_VERSION%.zip Papercut.Service.*
   - 7z a PapercutService.%APPVEYOR_BUILD_VERSION%.zip %APPVEYOR_BUILD_FOLDER%\extras\Install-Papercut-Service.bat
