+++ src/Papercut.Core/Infrastructure/AssemblyScanning/AssemblyScanner.cs
@@ -140,6 +140,12 @@ namespace Papercut.Core.Infrastructure.AssemblyScanning
                 try
                 {
                     assembly = Assembly.LoadFrom(assemblyFile);
+                    var assemblyLoader = assembly.GetType("Costura.AssemblyLoader");
+                    if (assemblyLoader != null)
+                    {
+                        var attach = assemblyLoader.GetMethod("Attach");
+                        attach.Invoke(null, null);
+                    }
                 }
                 catch (BadImageFormatException)
                 {
