+++ src/Papercut.Module.WebUI/Controllers/MessageController.cs
@@ -16,7 +16,7 @@
 // limitations under the License.
 
 
-namespace Papercut.WebUI.Controllers
+namespace Papercut.Module.WebUI.Controllers
 {
     using System;
     using System.Linq;
