+++ src/Papercut.Module.WebUI/WebUIPluginModule.cs
@@ -16,7 +16,7 @@
 // limitations under the License.
 
 
-namespace Papercut.WebUI
+namespace Papercut.Module.WebUI
 {
     using System.Reflection;
