+++ src/Papercut.Module.WebUI/assets/css/jquery-ui-1.10.4-smoothness.css
