+++ src/Papercut.Module.WebUI/assets/css/style.css
