+++ src/Papercut.Module.WebUI/assets/fonts/glyphicons-halflings-regular.woff2
