+++ src/Papercut.Module.WebUI/assets/images/github.png
