+++ src/Papercut.Module.WebUI/assets/images/papercut-logo.png
