+++ src/Papercut.Module.WebUI/assets/index.html
