+++ src/Papercut.Module.WebUI/assets/js/angular-1.3.8.js
