+++ src/Papercut.Module.WebUI/assets/js/controllers.js
