+++ src/Papercut.Module.WebUI/assets/js/iso88591_map.js
