+++ src/Papercut.Module.WebUI/assets/js/moment-2.8.4.js
