+++ src/Papercut.Module.WebUI/assets/js/punycode.js
