+++ src/Papercut.Module.WebUI/assets/js/sjis_map.js
