+++ src/Papercut.Module.WebUI/assets/js/strutil.js
