+++ src/Papercut.Module.WebUI/packages.config
@@ -11,4 +11,6 @@
   <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net40" />
   <package id="Rx-Core" version="2.2.5" targetFramework="net40" />
   <package id="Rx-Interfaces" version="2.2.5" targetFramework="net40" />
+  <package id="Costura.Fody" version="1.5.1" targetFramework="net40" developmentDependency="true" />
+  <package id="Fody" version="2.0.10" targetFramework="net40" developmentDependency="true" />
 </packages>
