+++ test/Papercut.Module.WebUI.Tests/packages.config
